requirements.txt
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
flask==2.3.2
werkzeug==2.3.6
pillow==10.0.0
numpy==1.24.3
